% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_distance.R
\name{est_distance}
\alias{est_distance}
\title{Linkage disequilibrium decay distance.}
\usage{
est_distance(ld, threshold = 0.3, k = 30)
}
\arguments{
\item{ld}{A data frame with two columns: \code{Bin} contains the distance 
between SNP pairs (or bin widths) and \code{R2} contains the estimated LD.}

\item{threshold}{The strength at which LD is considered decayed (default: 0.3).}

\item{k}{The number of knots to use for spline fitting (default: 30)}
}
\value{
The distance in units of the \code{Bin} column at which LD is estimated 
  to reach \code{threshold}.
}
\description{
Estimates the linkage disequilibrium (LD) decay distance by fitting a 
monotonically decreasing spline to estimates of pairwise LD.
}
\details{
This function estimates the LD decay distance by fitting a monotonically 
  decreasing spline to empirical estimates of LD. The units of \code{Bin} may 
  vary, depending on application. For example, local LD decay around an 
  individual SNP can be calculated where each row of \code{ld} is the measure 
  of LD between that SNP and a second SNP with the base-pair distance between 
  them. Alternatively, global LD decay could be calculated where LD measures 
  are aggregated into bins of 20,000 bp, for example.
}
\references{
Bustos-Korts \emph{et al.} (2019) Exome sequences and multi-environment 
  field trials elucidate the genetic basis of adaptation in barly. \emph{The Plant Journal}, 
  doi: 10.1111/tpj.14414.

Based on code from <https://gist.github.com/jnpaulson/c47f9bd3246f1121ad3a911e5f707355>.
}
