% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ibs}
\alias{ibs}
\title{Identity-by-state between genotypes.}
\usage{
ibs(X)
}
\arguments{
\item{X}{A numeric matrix of allelic dosages where rows represent different
genotypes and columns represent different markers. If \code{X} has rownames, 
these will be preserved in the output.}
}
\value{
A numeric matrix containing the proportion of shared alleles between
  each pair of genotypes.
}
\description{
Calculates the proportion of shared alleles between pairs of genotypes.
}
\details{
Each element of \code{X} should be in the interval \eqn{[0,2]}. 
  The matrix is pre-multiplied by one-half to correspond to equation 1 of 
  Bustos-Korts \emph{et al.} (2016).
}
\references{
Bustos-Korts \emph{et al.} (2016) Improvement of predictive 
  ability by uniform coverage of the target genetic space. \emph{G3}, 
  \strong{6}: 3733-47.
}
