\name{QGenTools-package}
\alias{QGenTools-package}
\alias{QGenTools}
\docType{package}
\title{
\packageTitle{QGenTools}
}
\description{
\packageDescription{QGenTools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{QGenTools}
\packageIndices{QGenTools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{QGenTools}

Maintainer: \packageMaintainer{QGenTools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
