% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r2}
\alias{r2}
\title{Correlation between loci.}
\usage{
r2(X)
}
\arguments{
\item{X}{An \eqn{n} x \eqn{m} numeric matrix of allelic dosages. The matrix must 
have column names for labelling correlations in the output.}
}
\value{
A dataframe with \eqn{m(m-1)/2} rows containing pairwise correlations
  between all possible pairs of markers in \code{X}.
}
\description{
Calculates the correlation (\eqn{r^2}) between marker loci within a population.
}
\details{
Elements of \code{X} should be in the interval \eqn{[0,2]}. This 
  computes the traditional \eqn{r^2} measure of linkage disequilibrium 
  between pairs of marker loci.
}
