% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r2s}
\alias{r2s}
\title{Correlation between loci.}
\usage{
r2s(X, st)
}
\arguments{
\item{X}{An \eqn{n} x \eqn{m} numeric matrix of allelic dosages. The matrix must 
have column names for labelling correlations in the output.}

\item{st}{An \eqn{n} x \eqn{K-1} numeric matrix of population structure 
covariates where \eqn{K} is the estimated number of subpopulations.}
}
\value{
A dataframe with \eqn{m(m-1)/2} rows containing pairwise correlations
  between all possible pairs of markers in \code{X}.
}
\description{
Calculates the structure corrected correlation (\eqn{r_S^2}) between marker 
  loci within a population.
}
\details{
Elements of \code{X} should be in the interval \eqn{[0,2]}. \code{st} 
  is usually produced by principal components analysis or contains admixture 
  proportions from a program like STRUCTURE or fastSTRUCTURE. This function 
  calculates the structure corrected correlation (\eqn{r_S^2}) between marker
  loci (equation 1 of the reference).
}
\references{
Mangin \emph{et al.} (2011) Novel measures of linkage disequilibrium
  that correct the bias due to population structure and relatedness. 
  \emph{Heredity}, \strong{108}\emph{(3)}: 285-91.
}
